requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
alembic==1.8.1
    # via
    #   -r requirements.in
    #   flask-migrate
bcrypt==4.0.1
    # via flask-bcrypt
blinker==1.5
    # via flask-mail
click==8.1.3
    # via flask
colorama==0.4.5
    # via click
email-validator==1.2.1
flask==2.2.1
    # via
    #   -r requirements.in
    #   flask-bcrypt
    #   flask-login
    #   flask-mail
    #   flask-migrate
    #   flask-sqlalchemy
    #   flask-wtf
flask-bcrypt==1.0.1
    # via -r requirements.in
flask-login==0.6.2
    # via -r requirements.in
flask-mail==0.9.1
    # via -r requirements.in
flask-migrate==3.1.0
    # via -r requirements.in
flask-sqlalchemy==2.5.1
    # via
    #   -r requirements.in
    #   flask-migrate
flask-wtf==1.0.1
    # via -r requirements.in
greenlet==1.1.3.post0
    # via sqlalchemy
gunicorn==20.1.0
itsdangerous==2.0.1
    # via
    #   -r requirements.in
    #   flask
    #   flask-wtf
jinja2==3.1.2
    # via flask
mako==1.2.3
    # via alembic
markupsafe==2.1.1
    # via
    #   jinja2
    #   mako
    #   werkzeug
    #   wtforms
mysqlclient==2.1.1
pillow==9.2.0
    # via -r requirements.in
python-dotenv==0.21.0
sqlalchemy==1.4.40
    # via
    #   -r requirements.in
    #   alembic
    #   flask-sqlalchemy
werkzeug==2.2.2
    # via
    #   flask
    #   flask-login
wtforms==3.0.1
    # via
    #   -r requirements.in
    #   flask-wtf